+++ libre_fastapi_jwt/__init__.py
@@ -1,5 +1,5 @@
 """FastAPI extension that provides JWT Auth support (secure, easy to use and lightweight)"""
 
-__version__ = "0.22.7"
+__version__ = "0.22.8"
 
 from .auth_jwt import AuthJWT, AuthJWTBearer